Move reader and writer into own files.
